processor/requirements.txt
psycopg2-binary==2.9.9
anthropic==0.25.1
python-slugify==8.0.4
pandas==2.2.1
python-dotenv==1.0.1